requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
cycler==0.10.0
Django==3.0.1
django-filter==2.2.0
djangorestframework==3.11.0
geographiclib==1.50
geopy==1.20.0
joblib==0.14.1
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
mysqlclient==1.4.6
numpy==1.18.0
pandas==0.25.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22
scipy==1.4.1
seaborn==0.9.0
selenium==3.141.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
